% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem-class.R
\name{is.csem}
\alias{is.csem}
\title{Test whether an object is a `csem` object}
\usage{
is.csem(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
`TRUE` if `x` inherits from `"csem"`, `FALSE` otherwise.
}
\description{
Test whether an object is a `csem` object
}
