% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csem-class.R
\name{is.csem}
\alias{is.csem}
\title{Test whether an object is a `csem` object}
\usage{
is.csem(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
`TRUE` if `x` inherits from `"csem"`, `FALSE` otherwise.
}
\description{
Test whether an object is a `csem` object
}
\examples{
set.seed(1)
d <- matrix(rbinom(80 * 15, 1, 0.5), nrow = 80)
fit <- suppressMessages(csem_gt(d, error_type = "absolute"))
is.csem(fit)
is.csem(list())

}
